cast to stop warnings

+++ src/plc/plc_codegen.c
@@ -142,7 +142,7 @@ pl_bytecode_builder *plc_codegen_stmt(plc_codegen_context *c, stmt *s) {
 				plc_codegen_stmt(c2, &(s->block.children[i]));
 			}
 			plc_codegen_context_add(c, c2);
-			pl_bytecode_builder_add(c->code, POPTO, {c->stacksize});
+			pl_bytecode_builder_add(c->code, POPTO, {(unsigned int)c->stacksize});
 			break;
 		}
 		case DEFFUNC: {
@@ -241,7 +241,7 @@ case OP_ ## op: \
 				pl_bytecode_builder_add(c->code, CALL, {(int)e->e.arity - 1});
 				break;
 			case OP_ARRAY:
-				pl_bytecode_builder_add(c->code, MAKEARRAY, {(int)e->e.arity});
+				pl_bytecode_builder_add(c->code, MAKEARRAY, {e->e.arity});
 				break;
 			case OP_EQUAL:
 				pl_bytecode_builder_add(c->code, EQUAL, {});
